Add files via upload
If you wish to download the pack theres the .mcpack files, and if you wish to edit the code you can too; for there are some things that may need to be corrected. you should be able to just download the files and install them. if you have any trouble, try Chat GPT, lol. This may be the last time i do this even if i have only done it twice.